+++ src/main/java/com/schedule/dao/StaffDaoImpl.java
@@ -1,6 +1,6 @@
 package com.schedule.dao;
 
-import com.schedule.domain.StaffDto;
+import com.schedule.dto.StaffDto;
 import org.apache.ibatis.session.SqlSession;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Repository;
