+++ src/main/java/com/schedule/dto/ShiftSchdDto.java
@@ -1,4 +1,4 @@
-package com.schedule.domain;
+package com.schedule.dto;
 
 public class ShiftSchdDto {
     String id; // 사번
